% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego_plot.R
\name{ego_plot}
\alias{ego_plot}
\title{Visualization of all ego networks in a network or list of ego networks.}
\usage{
ego_plot(networks, method = NULL, order = 1, ncol = 1, nrow = 1,
  pdf = FALSE, filename = NULL, gg_args = NULL, ...)
}
\arguments{
\item{networks}{Either an \emph{n} by \emph{n} adjacency matrix, a \code{igraph} object, or \code{network} object.}

\item{method}{Method to use to produce visualizations. Either 'igraph', 'network', or 'ggnet2'.}

\item{order}{0 to include only edges sent and received by the ego. 1 to include the ego's first-step neigborhood.}

\item{ncol}{Integer indicating the number of columns for displaying the visualizations.}

\item{nrow}{Integer indicating the number of rows for displaying the visualizations.}

\item{pdf}{Boolean. If TRUE, will output a pdf of visualiztions.}

\item{filename}{String indicating the file name to save PDF as.}

\item{gg_args}{Additional \code{ggplot2} layers called with the \code{+} operator.}

\item{...}{Additional arguments to pass to \code{method}.}
}
\value{
A list (of either \code{igraph} objects, \code{network} objects, or adjacency matrices) containing the ego networks for each node in the network`
}
\description{
Visualization of all ego networks in a network or list of ego networks.
}
\examples{
#Default settings
require(sna)
set.seed(6436)
net <- rgraph(20, 1, 0.25)
ego_plot(net)

#Pass an network object, return igraph visualizations
set.seed(6436)
net <- as.network(rgraph(20, 1, 0.25))
ego_plot(net, method = 'igraph')

#Pass additional arguments from visualization method
set.seed(6436)
net <- as.network(rgraph(20, 1, 0.25))
Use `?gplot` to see possible additional arguments
ego_plot(net, method = 'gplot', vertex.col = 'pink', edge.col = 'gray50')
Use ?igraph.plotting to see possible additional arguments
ego_plot(net, method = 'igraph', vertex.color = 'slategray', vertex.size = 15, vertex.label = NA)

#Pass additional `ggplot2` arguments
set.seed(6436)
net <- as.network(rgraph(20, 1, 0.25))
ego_plot(net, method = "ggnet2", color = "blue", gg_args = list(ggtitle("ggnet2 plot"), coord_equal()))

Save plots to a PDF file
set.seed(6436)
net <- as.network(rgraph(20, 1, 0.25))
ego_plot(net, method = "gplot", pdf = TRUE, filename = "ego networks")

See each ego's "star" network
set.seed(6436)
net <- as.network(rgraph(20, 1, 0.25))
ego_plot(net, order = 0)

# Don't pass arguments to gg_args unless method = "ggnet2"!
\dontrun{
set.seed(6436)
net <- as.network(rgraph(20, 1, 0.25))
ego_plot(net, method = "gplot", gg_args = list(ggtitle("No way!")))
}
}
