% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ego_networks.R
\name{find_ego_networks}
\alias{find_ego_networks}
\title{Fleixbly finds ego networks from sociocentric data in several formats.}
\usage{
find_ego_networks(network, order = 1, type = NULL, return = "matrix")
}
\arguments{
\item{network}{Either an \emph{n} by \emph{n} adjacency matrix, a \code{igraph} object, or \code{network} object.}

\item{order}{Either 0 (ego network contains only those edges that include the ego) or 1 (the ego's first step neighborhood).}

\item{type}{(Optional) The type of object passed to the \code{network} argument. Either "igraph", "matrix", or "network". If Not provided, will determine type from network provided.}

\item{return}{Determines what object type the function should return. Either "igraph", "matrix", or "network".}
}
\value{
A list (of either \code{igraph} objects, \code{network} objects, or adjacency matrices) containing the ego networks for each node in the network`
}
\description{
Fleixbly finds ego networks from sociocentric data in several formats.
}
\examples{
require(igraph)
set.seed(6436)
g <- sample_gnp(20, 0.25)
find_ego_network(g, type = "igraph", order = 1, return = "igraph")

require(sna)
set.seed(6436)
net <- as.network(rgraph(20, 1, 0.25))
# Accepts `network` objects, returns list of `igraph` objects.
find_ego_network(net, type = "network", order = 1, return = "igraph")
}
