% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_plot.R
\name{group_plot}
\alias{group_plot}
\title{Visualization of network the nodes arranged by group}
\usage{
group_plot(network, group, weight_within = 10, weight_between = 1,
  layout = "fr")
}
\arguments{
\item{network}{Either an \emph{n} by \emph{n} adjacency matrix, a \code{igraph} object, or \code{network} object.}

\item{group}{A vector of node attributes indicating group membership or a data frame of two columns, one with node names and one with group membership. Group membership should be of class character.}

\item{weight_within}{Determines how strongly groups cluster together. Default is 10.}

\item{weight_between}{Determines how weak between group ties cluster. Lower values spread groups further apart. Default is 1.}

\item{layout}{Either "fr" or "kk".}
}
\value{
A vector of range scores for nodes in network.
}
\description{
Visualization of network the nodes arranged by group
}
\examples{
# A sample vector of group membership
set.seed(385)
grp <- sample(1:3, 20, replace = TRUE)
# A sample network of the same order (20)
require(sna)
net <- rgraph(20, 1, 0.2)
group_plot(net, grp)
}
